+++ src/Misc/Commands.cs
@@ -88,14 +88,9 @@ namespace mpvnet
         public static void Open_DVD_Or_BD_Folder()
         {
             App.InvokeOnMainThread(new Action(() => {
-                using (var dialog = new FolderBrowserDialog())
-                {
-                    dialog.Description = "Select a DVD or Blu-ray folder.";
-                    dialog.ShowNewFolderButton = false;
-
+                using (var dialog = new BetterFolderBrowser())
                     if (dialog.ShowDialog() == DialogResult.OK)
                         Core.LoadDiskFolder(dialog.SelectedPath);
-                }
             }));
         }